feat: added default widgets class
